+++ back-end/src/main/java/com/tdd/backend/option/data/OptionResponse.java
@@ -1,4 +1,4 @@
-package com.tdd.backend.render.data;
+package com.tdd.backend.option.data;
 
 import java.util.List;
