+++ back-end/src/main/java/com/tdd/backend/option/render/SpecificCarOption.java
@@ -1,6 +1,8 @@
-package com.tdd.backend.render;
+package com.tdd.backend.option.render;
 
 import java.util.List;
+
+import com.tdd.backend.option.EntireOption;
 
 public enum SpecificCarOption {
 	IONIC6(List.of(
